+++ src/main/java/com/gitsh01/libertyvillagers/mixin/MeleeAttackGoalMixin.java
@@ -1,0 +1,29 @@
+package com.gitsh01.libertyvillagers.mixin;
+
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.ai.goal.MeleeAttackGoal;
+import net.minecraft.entity.mob.PathAwareEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import static com.gitsh01.libertyvillagers.LibertyVillagersMod.CONFIG;
+
+@Mixin(MeleeAttackGoal.class)
+public class MeleeAttackGoalMixin {
+
+    @Shadow
+    protected PathAwareEntity mob;
+
+    @Inject(method = "canStart", at = @At("HEAD"), cancellable = true)
+    public void canStart(CallbackInfoReturnable<Boolean> cir) {
+        if (this.mob.getType() == EntityType.IRON_GOLEM && CONFIG.golemsConfig.golemsAvoidWater) {
+            if (this.mob.getTarget() != null && this.mob.getTarget().isTouchingWater()) {
+                cir.setReturnValue(false);
+                cir.cancel();
+            }
+        }
+    }
+}